Add support for piping arbitrary user info from ASImageDownloader to the ASNetworkImageNodeDelegate
