+++ Source/Private/ASNetworkImageLoadInfo+Private.h
@@ -1,0 +1,22 @@
+//
+//  ASNetworkImageLoadInfo+Private.h
+//  AsyncDisplayKit
+//
+//  Created by Adlai on 1/30/18.
+//  Copyright © 2018 Facebook. All rights reserved.
+//
+
+#import <AsyncDisplayKit/ASNetworkImageLoadInfo.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface ASNetworkImageLoadInfo ()
+
+- (instancetype)initWithURL:(NSURL *)url
+                     source:(ASNetworkImageSource)source
+         downloadIdentifier:(nullable id)downloadIdentifier
+                   userInfo:(nullable id)userInfo;
+
+@end
+
+NS_ASSUME_NONNULL_END
